refactor user manager
