+++ src/modules/Modules.ts
@@ -1,8 +1,9 @@
 import _ from 'lodash';
 import { Quote, QuoteInfo } from 'src/database/quotes/QuotesManager';
 import { flagToEvent, getBiTInfo, lookupFlag } from 'ss-scene-flags';
-import { economyManager, quotesManager, userManager } from '../System';
+import { economyManager, quotesManager } from '../System';
 import { getOrCreateUserName } from '../util/UserUtils';
+import { getUser } from '../database/Users';
 
 /**
  * Represents a function that handles a given command or subset of commands
@@ -183,7 +184,7 @@ export const handleEconomyCommand: HandlerDelegate = async (
     const command = commandParts.shift();
     const [channelName] = metadata;
     const user = await getOrCreateUserName(sender);
-    const owner = userManager.getUser(channelName).userId;
+    const owner = getUser(channelName).userId;
     const currencyName = 'BiTcoins';
     if (command === 'money') {
         let target: string;
@@ -226,7 +227,7 @@ export const handleEconomyCommand: HandlerDelegate = async (
         const [receiver, amount] = commandParts;
         economyManager.addCurrency(
             await getOrCreateUserName(receiver),
-            userManager.getUser(channelName).userId,
+            getUser(channelName).userId,
             Number(amount),
         );
         return `gave ${receiver} ${amount} ${currencyName}`;
