+++ src/util/UserUtils.ts
@@ -1,11 +1,12 @@
-import { botApiClient, userManager } from '../System';
+import { botApiClient } from '../System';
+import { getUserByTwitchId, registerUserWithoutAuth } from '../database/Users';
 
 export const getOrCreateUserId = async (twitchId: string): Promise<number> => {
-    const user = userManager.getUserByTwitchId(twitchId).userId;
+    const user = getUserByTwitchId(twitchId).userId;
     if (user === -1) {
         const username = (await botApiClient.users.getUserById(twitchId))?.name;
         if (username === undefined) return -1;
-        return userManager.registerUserWithoutAuth(username, twitchId);
+        return registerUserWithoutAuth(username, twitchId);
     }
     return user;
 };
@@ -15,9 +16,9 @@ export const getOrCreateUserName = async (
 ): Promise<number> => {
     const twitchId = (await botApiClient.users.getUserByName(username))?.id;
     if (twitchId === undefined) return -1;
-    const user = userManager.getUserByTwitchId(twitchId).userId;
+    const user = getUserByTwitchId(twitchId).userId;
     if (user === -1) {
-        return userManager.registerUserWithoutAuth(username, twitchId);
+        return registerUserWithoutAuth(username, twitchId);
     }
     return user;
 };
